Updated the way animations play
Added a new variable to hold the number of animations currently playing.
This variable must be accessed in mutual exclusion, so there is a new
variable to act as lock. Both have been implemented as static class
members, so every instance accesses the same. Maybe these should be on
the gamemanager and access them through its instance? If this works, all
the animations should end before destroying the object.