New menu! And internal changes(boooring)
Added a main menu. Everything is obviously placeholder yet. The big
update is not visible, now the game doesn't rely on reloading the level
to advance. Instead, now it destroys the map, re-generates it, and
places the needed objects. PLUS! A little perk, now the spacebar toogles
the camera from whole map to player close up.